src/requirements.txt
SQLAlchemy~=2.0.23
werkzeug~=3.0.1
wtforms~=3.1.1
flask_login~=0.6
flask_wtf~=1.2.1
flask~=3.0.0
pillow~=10.2.0
email_validator~=2.1
requests~=2.31.0
pyotp
qrcode[pil]